requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# File Upload Support
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3

# Authentication & JWT
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# File Processing - CSV and Excel Support
pandas==2.1.4
openpyxl==3.1.2      # For .xlsx files
xlrd==2.0.1          # For .xls files